Seperate J2ObjCParser to its own file
